+++ chevah/keycert/ssh.py
@@ -119,7 +119,7 @@ def generate_ssh_key(options, open_method=None):
         key_type = options.key_type.lower()
 
         if not hasattr(options, 'key_file') or options.key_file is None:
-            options.key_file = 'id_%s' % (key_type)
+            options.key_file = u'id_%s' % (key_type)
 
         private_file = options.key_file
 
@@ -132,10 +132,7 @@ def generate_ssh_key(options, open_method=None):
 
         key = Key.generate(key_type=key_type, key_size=key_size)
 
-        private_file_path = local_filesystem.getEncodedPath(private_file)
-        public_file_path = local_filesystem.getEncodedPath(public_file)
-
-        with open_method(private_file_path, 'wb') as file_handler:
+        with open_method(_path(private_file), 'wb') as file_handler:
             _store_OpenSSH(key, private_file=file_handler)
 
         key_comment = None
@@ -145,7 +142,7 @@ def generate_ssh_key(options, open_method=None):
         else:
             message_comment = u'without a comment'
 
-        with open_method(public_file_path, 'wb') as file_handler:
+        with open_method(_path(public_file), 'wb') as file_handler:
             _store_OpenSSH(key, public_file=file_handler, comment=key_comment)
 
         message = (