+++ release-notes.rst
@@ -2,6 +2,19 @@ Relese notes for Chevah KeyCert
 ###############################
 
 
+1.3.4 - 20/04/2015
+==================
+
+* Update error message for small RSA key size.
+
+
+1.3.3 - 17/04/2015
+==================
+
+* Fall back to sha1 when sha256 is not available on OS to sign CSR.
+* Don't allow creating RSA keys less than 512.
+
+
 1.3.2 - 14/04/2015
 ==================